+++ internal/service/sagemaker/sweep.go
@@ -87,6 +87,11 @@ func RegisterSweepers() {
 		F:    sweepImages,
 	})
 
+	resource.AddTestSweepers("aws_sagemaker_mlflow_tracking_server", &resource.Sweeper{
+		Name: "aws_sagemaker_mlflow_tracking_server",
+		F:    sweepMlflowTrackingServers,
+	})
+
 	resource.AddTestSweepers("aws_sagemaker_model_package_group", &resource.Sweeper{
 		Name: "aws_sagemaker_model_package_group",
 		F:    sweepModelPackageGroups,
@@ -1067,6 +1072,47 @@ func sweepPipelines(region string) error {
 
 	if err := sweep.SweepOrchestrator(ctx, sweepResources); err != nil {
 		sweeperErrs = multierror.Append(sweeperErrs, fmt.Errorf("sweeping SageMaker Pipelines: %w", err))
+	}
+
+	return sweeperErrs.ErrorOrNil()
+}
+
+func sweepMlflowTrackingServers(region string) error {
+	ctx := sweep.Context(region)
+	client, err := sweep.SharedRegionalSweepClient(ctx, region)
+	if err != nil {
+		return fmt.Errorf("getting client: %s", err)
+	}
+	conn := client.SageMakerClient(ctx)
+
+	sweepResources := make([]sweep.Sweepable, 0)
+	var sweeperErrs *multierror.Error
+
+	pages := sagemaker.NewListMlflowTrackingServersPaginator(conn, &sagemaker.ListMlflowTrackingServersInput{})
+	for pages.HasMorePages() {
+		page, err := pages.NextPage(ctx)
+
+		if awsv2.SkipSweepError(err) {
+			log.Printf("[WARN] Skipping SageMaker Mlflow Tracking Server sweep for %s: %s", region, err)
+			return sweeperErrs.ErrorOrNil()
+		}
+		if err != nil {
+			sweeperErrs = multierror.Append(sweeperErrs, fmt.Errorf("retrieving SageMaker Mlflow Tracking Servers: %w", err))
+		}
+
+		for _, project := range page.TrackingServerSummaries {
+			name := aws.ToString(project.TrackingServerName)
+
+			r := resourceMlflowTrackingServer()
+			d := r.Data(nil)
+			d.SetId(name)
+
+			sweepResources = append(sweepResources, sweep.NewSweepResource(r, d, client))
+		}
+	}
+
+	if err := sweep.SweepOrchestrator(ctx, sweepResources); err != nil {
+		sweeperErrs = multierror.Append(sweeperErrs, fmt.Errorf("sweeping SageMaker Mlflow Tracking Servers: %w", err))
 	}
 
 	return sweeperErrs.ErrorOrNil()